2nd wave of fixes

+++ index.js
@@ -6,8 +6,9 @@ const {
   deleteFromExpensesFile,
   updateExpensesFile,
   getExpensesSummary,
+  checkIdExists,
 } = require("./src/expenses");
-const { setbudget } = require("./src/budgets");
+const { setBudget } = require("./src/budgets");
 const exportToCSV = require("./src/utils/exportCSV");
 
 const program = new Command();
@@ -21,6 +22,17 @@ program
   .requiredOption("-c, --category <category>", "Category name")
   .action(async (options) => {
     const { description, amount, category } = options;
+
+    if (!description || !amount || !category) {
+      console.log("Please provide both description, amount and category");
+      return;
+    }
+
+    if (isNaN(amount) || amount <= 0) {
+      console.log("Amount must be a number greater than 0");
+      return;
+    }
+
     await writeToExpenseFile({ description, amount, category });
   });
 
@@ -53,7 +65,8 @@ program
   .requiredOption("-i, --id <id>", "Expense ID")
   .action(async (options) => {
     const { id } = options;
-    await deleteFromExpensesFile(id);
+    if (await checkIdExists(id)) await deleteFromExpensesFile(id);
+    else console.log(`Expense with ID ${id} does not exist`);
   });
 
 // Update command
@@ -65,17 +78,18 @@ program
   .option("-a --amount <amount>", "Expense amount")
   .action(async (options) => {
     const { id, description, amount } = options;
+    if (await checkIdExists(id)) {
+      if (!description && !amount) {
+        console.log("Please provide either description or amount to update");
+        return;
+      }
 
-    if (!description && !amount) {
-      console.log("Please provide either description or amount to update");
-      return;
-    }
+      const data = {};
+      if (description) data.description = description;
+      if (amount) data.amount = amount;
 
-    const data = {};
-    if (description) data.description = description;
-    if (amount) data.amount = amount;
-
-    await updateExpensesFile(id, data);
+      await updateExpensesFile(id, data);
+    } else console.log(`Expense with ID ${id} does not exist`);
   });
 
 // Summary command
@@ -85,8 +99,26 @@ program
   .option("-m, --month <month>", "Filter by month")
   .action(async (options) => {
     const { month } = options;
+    let commandMonth = "";
+    if (month) {
+      let monthMap = new Map([
+        [1, "Jan"],
+        [2, "Feb"],
+        [3, "Mar"],
+        [4, "Apr"],
+        [5, "May"],
+        [6, "Jun"],
+        [7, "Jul"],
+        [8, "Aug"],
+        [9, "Sep"],
+        [10, "Oct"],
+        [11, "Nov"],
+        [12, "Dec"],
+      ]);
+      commandMonth = monthMap.get(+month);
+    }
     const expenses = await readFromExpensesFile();
-    await getExpensesSummary(expenses, month);
+    await getExpensesSummary(expenses, commandMonth);
   });
 
 // Set budget command
@@ -97,7 +129,7 @@ program
   .requiredOption("-a, --amount <amount>", "Budget amount")
   .action(async (options) => {
     const { month, amount } = options;
-    await setbudget(month, amount);
+    await setBudget(month, amount);
   });
 
 // export command
